tidy(ui): move new dnd impl into features/dnd
