+++ invokeai/frontend/web/src/features/controlLayers/components/CanvasDropArea.tsx
@@ -1,6 +1,6 @@
 import { Grid, GridItem } from '@invoke-ai/ui-library';
-import { Dnd } from 'features/dnd2/dnd';
-import { DndDropTarget } from 'features/dnd2/DndDropTarget';
+import { Dnd } from 'features/dnd/dnd';
+import { DndDropTarget } from 'features/dnd/DndDropTarget';
 import { useImageViewer } from 'features/gallery/components/ImageViewer/useImageViewer';
 import { memo } from 'react';
 import { useTranslation } from 'react-i18next';
