+++ invokeai/frontend/web/src/features/dnd/DndImageIcon.tsx
